brought audioPlayer into branch
Minor twiddling with the delay effect, experimenting in SimpleDelay.process()

+++ DelayEffects/SimpleDelay.cpp
@@ -13,6 +13,6 @@
 double SimpleDelay::process(double inputSample)
 {
 		delaySample(inputSample*.707);
-		return (getInterpolatedOut(currentDelayIndex) + inputSample);
+		return (getInterpolatedOut(double(currentDelayIndex)) + inputSample);
 //		return inputSample;
 }
